refactor(workbench): W3 将 assistant 真抽入 @sailfish/workbench-assistant
descriptor/prompt/agent-tools/AssistantWorkbench 迁入包内；registry 与
electron tools 改包名 import；src 保留兼容 re-export。artifact/AiPanel
仍依赖 desktop（P2）。
