+++ docs/workbench-monorepo-design.md
@@ -548,7 +548,7 @@ apps/desktop/                               @sailfish/core-team
 | **Bootstrap 装配** | 启动连 MCP、校验 skills | ⚠️ | 端到端验证；MCP 就绪信号；skills 缺失告警可产品化 |
 | **P-1 shared-types** | **物理迁入**真包 + 全仓 import | ⚠️ | ✅ 物理迁入 `packages/shared-types`；`shared/types` 为兼容 re-export；全仓改 `@sailfish/shared-types` 可渐进 |
 | **Workspace** | pnpm（决策）或过渡 npm workspaces | ⚠️ | ✅ npm workspaces 已启用（W1）；❌ 尚未切 pnpm（W5） |
-| **P0 assistant 抽包** | 物理迁 `workbench-assistant` + 构建冒烟 | ⚠️ | ❌ 迁 Vue/descriptor 真源；❌ 验证 HMR / electron-builder |
+| **P0 assistant 抽包** | 物理迁 `workbench-assistant` + 构建冒烟 | ⚠️ | ✅ descriptor/prompt/agent-tools/AssistantWorkbench 真源在包内；`src/` 薄 re-export；artifact/AiPanel 仍 desktop（P2）；electron-builder 全量冒烟待做 |
 | **P1 全台 + SDK** | local/ssh/companion + sdk 真包；region 渲染器 | ⚠️ | SDK 仍 re-export；❌ `iframe-url` / `data-table`；❌ `apps/desktop` 拆分 |
 | **P2 AiPanel 下沉** | 对话区可被业务台复用 | ⏸❌ | 见 `src/components/AIPANEL_SPEC.md`；高风险，单独排期 |
 | **P3 useAgentMode** | 原语进 SDK | ⏸❌ | 后置 |
@@ -561,7 +561,7 @@ apps/desktop/                               @sailfish/core-team
 
 - [x] **W1** 启用 **npm workspaces**（不换 pnpm）：`package.json#workspaces` + `@sailfish/*` 依赖声明；`npm install` / 类型检查冒烟（2026-07-13）  
 - [x] **W2** **P-1 物理迁码**：`shared/types` → `packages/shared-types`；保留 `@shared/types` alias（2026-07-13）  
-- [ ] **W3** **P0**：`assistant` 真抽包（descriptor + prompt + AssistantWorkbench）+ desktop import 包名  
+- [x] **W3** **P0**：`assistant` 真抽包（descriptor + prompt + agent-tools + AssistantWorkbench）+ desktop/registry/tools 改包名 import（2026-07-13）；artifact 与 AiPanel 仍留 desktop 
 - [ ] **W4** 样例业务台（哪怕无对话：descriptor + 假 MCP）验证 bootstrap  
 - [ ] **W5** 评估切 **pnpm**（与设计决策 #2 对齐）或文档改为「允许 npm workspaces 过渡」  
 - [ ] **W6** P1 余量：local/ssh/companion 真抽 + SDK 去 re-export
