+++ packages/workbench-assistant/package.json
@@ -2,11 +2,20 @@
   "name": "@sailfish/workbench-assistant",
   "version": "0.1.0",
   "private": true,
-  "description": "独立助手工作台包（骨架：descriptor + prompt）",
+  "description": "独立助手工作台（descriptor / prompt / AssistantWorkbench）",
   "type": "module",
   "main": "./src/index.ts",
   "types": "./src/index.ts",
   "exports": {
-    ".": "./src/index.ts"
+    ".": "./src/index.ts",
+    "./descriptor": "./src/descriptor.ts",
+    "./prompt": "./src/prompt.ts",
+    "./agent-tools": "./src/agent-tools.ts"
+  },
+  "peerDependencies": {
+    "vue": "^3.4.0"
+  },
+  "dependencies": {
+    "@sailfish/shared-types": "*"
   }
 }
