+++ src/workbench/assistant/descriptor.ts
@@ -1,10 +1,2 @@
-import AssistantWorkbench from '../../components/workbench/AssistantWorkbench.vue'
-import type { WorkbenchDescriptor } from '../types'
-import { AGENT_PROMPT, shouldInjectAgentPrompt } from './prompt'
-
-export const descriptor: WorkbenchDescriptor = {
-  kind: 'assistant',
-  renderer: AssistantWorkbench,
-  availableInSteam: false,
-  agentPrompt: (tab) => (shouldInjectAgentPrompt(tab) ? AGENT_PROMPT : undefined),
-}
+/** @deprecated 兼容 re-export；真相源 `@sailfish/workbench-assistant/descriptor` */
+export { descriptor } from '@sailfish/workbench-assistant/descriptor'
