+++ tsconfig.json
@@ -24,6 +24,7 @@
       "@sailfish/shared-types": ["./packages/shared-types/src/index.ts"],
       "@sailfish/workbench-sdk": ["./packages/workbench-sdk/src/index.ts"],
       "@sailfish/workbench-assistant": ["./packages/workbench-assistant/src/index.ts"],
+      "@sailfish/workbench-assistant/*": ["./packages/workbench-assistant/src/*"],
       "@sailfish/workbench-local": ["./packages/workbench-local/src/index.ts"],
       "@sailfish/workbench-ssh": ["./packages/workbench-ssh/src/index.ts"],
       "@sailfish/workbench-companion": ["./packages/workbench-companion/src/index.ts"]
@@ -34,7 +35,9 @@
     "src/**/*.tsx",
     "src/**/*.vue",
     "shared/**/*.ts",
-    "packages/shared-types/src/**/*.ts"
+    "packages/shared-types/src/**/*.ts",
+    "packages/workbench-assistant/src/**/*.ts",
+    "packages/workbench-assistant/src/**/*.vue"
   ],
   "exclude": ["src/**/__tests__/**", "src/**/*.test.ts"]
 }
